% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-distance.R
\name{calc_dist_matrix}
\alias{calc_dist_matrix}
\title{Calculate distance matrix in parallel}
\usage{
calc_dist_matrix(x, output_dir = ".", pb = NULL)
}
\arguments{
\item{x}{simple feature object including at least one column called
\code{ISO3_CODE}. See \code{?sf}}

\item{output_dir}{path to store distance matrix RDS files.
Default is work dir}

\item{pb}{progress bar. See \code{?progress::progress_bar}.
Default is NULL}
}
\value{
TRUE if success FALSE if could not create the distance matrix
}
\description{
Calculate distance matrix in parallel
}
